Create DashboardDetailsHeader, DashboardDetailsFooter, and update DashboardDetailsApp to render the full main content twice, once using WidgetContextRenderer, the other one manually.
