--- assets/js/components/dashboard-details/dashboard-details-module.js
@@ -1,34 +1,0 @@
-/**
- * DashboardDetailsModule component.
- *
- * Site Kit by Google, Copyright 2021 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * WordPress dependencies
- */
-import { withFilters } from '@wordpress/components';
-import { Component } from '@wordpress/element';
-
-/**
- * A single module. Keeps track of its own active state and settings.
- */
-class DashboardDetailsModule extends Component {
-	render() {
-		return null;
-	}
-}
-
-export default withFilters( 'googlesitekit.DashboardDetailsModule' )( DashboardDetailsModule );
